move ssg into a separate crate
