+++ Cargo.toml
@@ -29,6 +29,7 @@ members = [
     "packages/hot-reload",
     "packages/fullstack",
     "packages/server-macro",
+    "packages/static-generation",
     "packages/fullstack/examples/axum-hello-world",
     "packages/fullstack/examples/axum-router",
     "packages/fullstack/examples/axum-streaming",
@@ -76,8 +77,8 @@ dioxus-cli-config = { path = "packages/cli-config", version = "0.5.0", default-f
 generational-box = { path = "packages/generational-box", version = "0.5.0" }
 dioxus-hot-reload = { path = "packages/hot-reload", version = "0.5.0" }
 dioxus-fullstack = { path = "packages/fullstack", version = "0.5.0" }
+dioxus-static-site-generation = { path = "packages/static-generation", version = "0.5.0" }
 dioxus_server_macro = { path = "packages/server-macro", version = "0.5.0", default-features = false }
-dioxus-ext = { path = "packages/extension" }
 tracing = "0.1.37"
 tracing-futures = "0.2.5"
 toml = "0.8"
@@ -119,9 +120,6 @@ isnta = "1.36.1"
 
 # speed up some macros by optimizing them
 [profile.dev.package.insta]
-opt-level = 3
-
-[profile.dev.package.similar]
 opt-level = 3
 
 [profile.dev.package.dioxus-core-macro]
