+++ packages/fullstack/src/render.rs
@@ -211,18 +211,10 @@ impl SSRState {
 }
 
 /// The template that wraps the body of the HTML for a fullstack page. This template contains the data needed to hydrate server functions that were run on the server.
+#[derive(Default)]
 pub struct FullstackHTMLTemplate {
     cfg: ServeConfig,
     server_context: DioxusServerContext,
-}
-
-impl Default for FullstackHTMLTemplate {
-    fn default() -> Self {
-        Self {
-            cfg: ServeConfig::default(),
-            server_context: DioxusServerContext::default(),
-        }
-    }
 }
 
 impl FullstackHTMLTemplate {
