--- packages/fullstack/src/static_generation/mod.rs
@@ -1,3 +1,0 @@
-mod config;
-pub use config::*;
-pub mod launch;
