+++ packages/ssr/src/cache.rs
@@ -6,7 +6,6 @@ use crate::renderer::{str_truthy, BOOL_ATTRS};
 #[derive(Debug)]
 pub struct StringCache {
     pub segments: Vec<Segment>,
-    pub template: Template,
 }
 
 #[derive(Default)]
@@ -56,7 +55,6 @@ impl StringCache {
 
         Ok(Self {
             segments: chain.segments,
-            template: template.template.get(),
         })
     }
