+++ packages/static-generation/.gitignore
@@ -1,0 +1,1 @@
+target
