+++ packages/static-generation/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "dioxus-static-site-generation"
+authors = ["Evan Almloff"]
+version = { workspace = true }
+edition = "2021"
+description = "Static site generation for Dioxus"
+license = "MIT OR Apache-2.0"
+repository = "https://github.com/DioxusLabs/dioxus/"
+homepage = "https://dioxuslabs.com"
+keywords = ["ui", "gui", "react", "ssg"]
+resolver = "2"
+
+[dependencies]
+dioxus-fullstack = { workspace = true }
+dioxus-lib.workspace = true
+dioxus-router = { workspace = true }
+dioxus-ssr = { workspace = true, features = ["incremental"], optional = true }
+dioxus-web = { workspace = true, features = ["hydrate"], optional = true }
+tokio = { workspace = true, optional = true }
+http = { workspace = true, optional = true }
+
+[dev-dependencies]
+dioxus = { workspace = true }
+
+[features]
+default = []
+site-generation = ["dioxus-fullstack/server", "dioxus-router/ssr", "dioxus-ssr", "tokio", "http"]
+web = ["dioxus-fullstack/web", "dioxus-router/web", "dioxus-web"]
