+++ packages/static-generation/src/lib.rs
@@ -1,0 +1,6 @@
+mod config;
+pub use config::*;
+pub mod launch;
+
+#[cfg(feature = "site-generation")]
+pub(crate) mod ssg;
